requirements.txt
Flask
psycopg2-binary
python-dotenv
flasgger
requests
gunicorn